# optimise bitmaps