@lemire 2021-01-19
Thanks for the PR.

I am going to take a critical stance here, however... because the Go compiler will definitively compile these functions to the same output:

```Go
func f1(x uint) uint {
    return x / 64
}

func f2(x uint) uint {
    return x >> 6
}
```

So we need to examine these changes we care.