@lemire 2021-01-19
> On it boss

Note that I am not trying to be difficult.

The approach you are taking is the one I took for my code contributions to https://github.com/willf/bitset/blob/master/bitset.go

So I wrote many of the same lines of code before!!!

The Java version of Roaring is also filled with shifts and masks.


But I have gained a bit of trust in the Go compiler since then, and I would write the code differently.