@seiflotfy 2021-01-19
> @seiflotfy
> 
> This is a well motivated PR and there are almost certainly some real gains. We want to go forward with this PR.
> 
> Before we do, please consider the following:
> 
> 1. I can do shift and masks in my head and know exactly what they mean, and presumably, you can too. For many other programmers, however, it makes the code hard to read. I submit to you that we should always try to favour code that is easy to read, unless we can demonstrate the benefits of our "hard to read code".

I agree I prefer readable code over non-readable. So I will try to keep it that way.

> 2. The Go compiler is definitively smart enough to turn `x/64` and `x%64` into fast instructions when `x` is unsigned. When `x` is signed... well, your current PR changes the semantics by forcing the assumption that the variable is unsigned. We might as well go one step further and actually make the variable unsigned and then rely on the compiler. I submit to you that all cases except for my next point could be solved by turning the variables into unsigned integers. That is, I think you can achieve the very same result just by switching the signed integers to unsigned integers.  If I am wrong, it would be interesting to know why and document it.

I will try to switch to unsigned integers and will document the results.

> 3. At one precise point you turn a branchless routine into a branchy routine. I see you have reverted that. If you believe this is faster, then one should be really careful about the benchmark because if you use a benchmark where the branch is easily predicted, you will get artificially good results. You have to include a benchmark where the branch is really hard to predict... and in this case, it means that `iadd` should be sometimes adding a value that is already present, and sometimes not. You can look at the assembly of the current `iadd` function: https://godbolt.org/z/38qhjr As you can see, it is really, really cheap.

I want to tackle the shifting problem before getting deeper into this so I will do this in a separate branch later :)

> 4. Please have a look at the assembly of the functions: if the assembly is identical, then there is no need for benchmarking. And if the instruction count is reduced, then we also do not need benchmarking. That is, when tweaking purely computational code the way you do, I believe it is best to look at the assembly and just 'count the instructions'. That is, you turn a division into a shift, if you should be able to see it in the assembly, and then we don't need to benchmark, we know that a shift is cheaper. One can use godbolt or other more direct techniques to examine the assembly.

I will look into the assembly functions and try to document my findings as part of the PR :)