@seiflotfy 2021-01-25
is there a possibility of the argument `i` in `func (bc *bitmapContainer) NextSetBit(i int) int` being negative? I don't think so