@lemire 2021-01-25
@seiflotfy 

It can return a negative value:

```Go
func (bc *bitmapContainer) NextSetBit(i int) int {
	x := i / 64
	if x >= len(bc.bitmap) {
		return -1
	}
	w := bc.bitmap[x]
	w = w >> uint(i%64)
	if w != 0 {
		return i + countTrailingZeros(w)
	}
	x++
	for ; x < len(bc.bitmap); x++ {
		if bc.bitmap[x] != 0 {
			return (x * 64) + countTrailingZeros(bc.bitmap[x])
		}
	}
	return -1
}
```

But if you pass it a negative value (as an input) then you almost surely have a bug. (This could be explicitly documented.)

So yeah, the input could be made unsigned.