@seiflotfy 2021-01-25
with a little change
```go
func (bc *bitmapContainer) NextSetBit(i int) int {
	var (
		x      = uint(i) / 64
		length = uint(len(bc.bitmap))
	)

	if x >= length {
		return -1
	}

	w := bc.bitmap[x]
	w = w >> (uint(i) % 64)
	if w != 0 {
		return i + countTrailingZeros(w)
	}
	x++

	for ; x < length; x++ {
		if bc.bitmap[x] != 0 {
			return int(x*64) + countTrailingZeros(bc.bitmap[x])
		}
	}

	return -1
}
```
I was able to improve the iterator benchmarks

```

name                                                           old time/op    new time/op    delta
ShortIteratorNextBitmap/init_size_1_shortIterator_next-12         561ns ± 1%     312ns ± 0%  -44.39%  (p=0.002 n=10+4)
ShortIteratorNextBitmap/init_size_650_shortIterator_next-12      5.02µs ± 0%    4.35µs ± 1%  -13.36%  (p=0.001 n=9+5)
ShortIteratorNextBitmap/init_size_6500_shortIterator_next-12     45.0µs ± 0%    40.6µs ± 1%   -9.81%  (p=0.001 n=10+5)
ShortIteratorNextBitmap/init_size_65535_shortIterator_next-12     449µs ± 0%     409µs ± 1%   -8.91%  (p=0.001 n=9+5)

name                                                           old alloc/op   new alloc/op   delta
ShortIteratorNextBitmap/init_size_1_shortIterator_next-12         16.0B ± 0%     16.0B ± 0%     ~     (all equal)
ShortIteratorNextBitmap/init_size_650_shortIterator_next-12       16.0B ± 0%     16.0B ± 0%     ~     (all equal)
ShortIteratorNextBitmap/init_size_6500_shortIterator_next-12      16.0B ± 0%     16.0B ± 0%     ~     (all equal)
ShortIteratorNextBitmap/init_size_65535_shortIterator_next-12     16.0B ± 0%     16.0B ± 0%     ~     (all equal)

name                                                           old allocs/op  new allocs/op  delta
ShortIteratorNextBitmap/init_size_1_shortIterator_next-12          1.00 ± 0%      1.00 ± 0%     ~     (all equal)
ShortIteratorNextBitmap/init_size_650_shortIterator_next-12        1.00 ± 0%      1.00 ± 0%     ~     (all equal)
ShortIteratorNextBitmap/init_size_6500_shortIterator_next-12       1.00 ± 0%      1.00 ± 0%     ~     (all equal)
ShortIteratorNextBitmap/init_size_65535_shortIterator_next-12      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```

A cleaner way is to change the input to unsigned as you mentioned... Which one do you prefer?