@lemire 2021-01-25
> A cleaner way is to change the input to unsigned as you mentioned... Which one do you prefer?


I submit to you that your latest submitted code is *more* readable than the one with shifts.

If a variable is unsigned, the Go compiler is *smart* enough to replace a division by 64 with a shift. 

For signed values, it may not be able to do and still abide by the spec. Now, it could try to find out that the variable is always non-negative but I think we agree that the Go compiler is not smart enough to do that.