@seiflotfy 2021-01-25
Updated the benchmarks, 

```
name                                                                  old time/op    new time/op    delta
ShortIteratorAdvanceArray/init_size_1_shortIterator_advance-12          41.8ns ± 0%    40.3ns ± 0%   -3.48%  (p=0.016 n=4+5)
ShortIteratorAdvanceArray/init_size_650_shortIterator_advance-12        77.2ns ± 1%    76.1ns ± 2%     ~     (p=0.659 n=5+5)
ShortIteratorAdvanceArray/init_size_6500_shortIterator_advance-12       86.8ns ± 2%    89.4ns ± 5%     ~     (p=0.056 n=5+5)
ShortIteratorAdvanceArray/init_size_65535_shortIterator_advance-12      90.2ns ± 0%    90.0ns ± 2%     ~     (p=0.651 n=5+5)
ShortIteratorNextArray/init_size_1_shortIterator_next-12                37.9ns ± 1%    37.4ns ± 6%     ~     (p=0.135 n=5+5)
ShortIteratorNextArray/init_size_650_shortIterator_next-12              2.35µs ± 3%    2.23µs ± 1%   -5.29%  (p=0.008 n=5+5)
ShortIteratorNextArray/init_size_6500_shortIterator_next-12             22.2µs ± 1%    22.0µs ± 3%     ~     (p=0.690 n=5+5)
ShortIteratorNextArray/init_size_65535_shortIterator_next-12             226µs ± 1%     230µs ± 9%     ~     (p=0.841 n=5+5)
ShortIteratorAdvanceBitmap/init_size_1_shortIterator_advance-12         37.9ns ± 1%    36.2ns ± 1%   -4.49%  (p=0.008 n=5+5)
ShortIteratorAdvanceBitmap/init_size_650_shortIterator_advance-12       42.1ns ± 1%    38.9ns ± 0%   -7.74%  (p=0.008 n=5+5)
ShortIteratorAdvanceBitmap/init_size_6500_shortIterator_advance-12      44.1ns ± 3%    38.8ns ± 1%  -12.14%  (p=0.008 n=5+5)
ShortIteratorAdvanceBitmap/init_size_65535_shortIterator_advance-12     44.9ns ±17%    38.9ns ± 0%  -13.36%  (p=0.008 n=5+5)
ShortIteratorNextBitmap/init_size_1_shortIterator_next-12                582ns ± 2%     313ns ± 1%  -46.25%  (p=0.008 n=5+5)
ShortIteratorNextBitmap/init_size_650_shortIterator_next-12             5.14µs ± 1%    4.36µs ± 1%  -15.24%  (p=0.008 n=5+5)
ShortIteratorNextBitmap/init_size_6500_shortIterator_next-12            45.8µs ± 2%    40.8µs ± 1%  -10.93%  (p=0.008 n=5+5)
ShortIteratorNextBitmap/init_size_65535_shortIterator_next-12            455µs ± 0%     416µs ± 4%   -8.45%  (p=0.008 n=5+5)
ShortIteratorAdvanceRuntime/init_size_1_shortIterator_advance-12        43.8ns ± 1%    42.2ns ± 2%   -3.47%  (p=0.008 n=5+5)
ShortIteratorAdvanceRuntime/init_size_650_shortIterator_advance-12      49.2ns ± 1%    47.5ns ± 1%   -3.45%  (p=0.008 n=5+5)
ShortIteratorAdvanceRuntime/init_size_6500_shortIterator_advance-12     49.2ns ± 1%    47.6ns ± 1%   -3.21%  (p=0.008 n=5+5)
ShortIteratorAdvanceRuntime/init_size_65535_shortIterator_advance-12    49.0ns ± 1%    47.9ns ± 1%   -2.16%  (p=0.008 n=5+5)
ShortIteratorNextRuntime/init_size_1_shortIterator_next-12              40.4ns ± 0%    39.2ns ± 1%   -2.97%  (p=0.008 n=5+5)
ShortIteratorNextRuntime/init_size_650_shortIterator_next-12            3.60µs ± 2%    3.52µs ± 3%     ~     (p=0.095 n=5+5)
ShortIteratorNextRuntime/init_size_6500_shortIterator_next-12           35.1µs ± 1%    34.2µs ± 0%   -2.71%  (p=0.008 n=5+5)
ShortIteratorNextRuntime/init_size_65535_shortIterator_next-12           352µs ± 1%     345µs ± 0%   -2.04%  (p=0.008 n=5+5)
```

Its a minimal change. I would suggest that if you would merge this. I would continue looking for other places where things get optimised due to the compiler being not smart enough, making this initiative a series of small changes and improvements.